+++ source/extensions/isaacsim.core.prims/python/impl/sdf_shape_prim.py
@@ -161,11 +161,6 @@ class SdfShapePrim(GeometryPrim):
         """
         return self._physics_view is not None
 
-    def _invalidate_physics_handle_callback(self, event):
-        if event.type == int(omni.timeline.TimelineEventType.STOP):
-            self._physics_view = None
-        return
-
     def initialize(self, physics_sim_view: omni.physics.tensors.SimulationView = None) -> None:
         """Create a physics simulation view if not passed and creates a sdf shape view in physX.