+++ source/extensions/isaacsim.examples.ui/isaacsim/examples/ui/extension.py
@@ -395,19 +395,13 @@ class Extension(omni.ext.IExt):
                 ui.Spacer()
 
     def toggle_app_step(self, val=None):
-        print("You've cliked time_series_plot_data:", val)
-        if val:
-            if not self._app_event_sub:
-                self._app_event_sub = carb.eventdispatcher.get_eventdispatcher().observe_event(
-                    event_name=omni.kit.app.GLOBAL_EVENT_UPDATE,
-                    on_event=self._on_app_step,
-                    observer_name="isaacsim.examples.ui.Extension._on_app_step",
-                )
-                # self._app_event_sub = (
-                #     omni.kit.app.get_app().get_update_event_stream().create_subscription_to_pop(self._on_app_step)
-                # )
-            else:
-                self._app_event_sub = None
+        print("You've clicked time_series_plot_data:", val)
+        if val and self._app_event_sub is None:
+            self._app_event_sub = carb.eventdispatcher.get_eventdispatcher().observe_event(
+                event_name=omni.kit.app.GLOBAL_EVENT_UPDATE,
+                on_event=self._on_app_step,
+                observer_name="isaacsim.examples.ui.Extension._on_app_step",
+            )
         else:
             self._app_event_sub = None
 
@@ -421,19 +415,13 @@ class Extension(omni.ext.IExt):
         self._models["timeseries_plot"].set_data(*self._plot_data)
 
     def toggle_app_step_1(self, val=None):
-        print("You've cliked time_series_plot_data:", val)
-        if val:
-            if not self._app_event_sub:
-                self._app_event_sub = carb.eventdispatcher.get_eventdispatcher().observe_event(
-                    event_name=omni.kit.app.GLOBAL_EVENT_UPDATE,
-                    on_event=self._on_app_step_1,
-                    observer_name="isaacsim.examples.ui.Extension._on_app_step_1",
-                )
-                # self._app_event_sub = (
-                #     omni.kit.app.get_app().get_update_event_stream().create_subscription_to_pop(self._on_app_step_1)
-                # )
-            else:
-                self._app_event_sub = None
+        print("You've clicked time_series_plot_data:", val)
+        if val and self._app_event_sub is None:
+            self._app_event_sub = carb.eventdispatcher.get_eventdispatcher().observe_event(
+                event_name=omni.kit.app.GLOBAL_EVENT_UPDATE,
+                on_event=self._on_app_step_1,
+                observer_name="isaacsim.examples.ui.Extension._on_app_step_1",
+            )
         else:
             self._app_event_sub = None
 
@@ -447,16 +435,13 @@ class Extension(omni.ext.IExt):
         self._models["timeseries_plot_hist"].set_data(*self._plot_data)
 
     def toggle_app_step_2(self, val=None):
-        print("You've cliked time_series_plot_data:", val)
-        if val:
-            if not self._app_event_sub:
-                self._app_event_sub = carb.eventdispatcher.get_eventdispatcher().observe_event(
-                    event_name=omni.kit.app.GLOBAL_EVENT_UPDATE,
-                    on_event=self._on_app_step_2,
-                    observer_name="isaacsim.examples.ui.Extension._on_app_step_2",
-                )
-            else:
-                self._app_event_sub = None
+        print("You've clicked time_series_plot_data:", val)
+        if val and self._app_event_sub is None:
+            self._app_event_sub = carb.eventdispatcher.get_eventdispatcher().observe_event(
+                event_name=omni.kit.app.GLOBAL_EVENT_UPDATE,
+                on_event=self._on_app_step_2,
+                observer_name="isaacsim.examples.ui.Extension._on_app_step_2",
+            )
         else:
             self._app_event_sub = None