+++ source/extensions/isaacsim.replicator.behavior/isaacsim/replicator/behavior/behaviors/example_custom_event_behavior.py
@@ -84,8 +84,6 @@ class ExampleCustomEventBehavior(BehaviorScript):
         """Called when the script is unassigned from a prim."""
         # Unsubscribe from the event stream
         self._reset()
-
-        self._event_sub.reset()
         self._event_sub = None
 
         # Exposed variables should be removed if the script is no longer assigned to the prim