+++ source/standalone_examples/testing/isaacsim.simulation_app/test_viewport_ready.py
@@ -71,4 +71,5 @@ try:
 
 finally:
     # Cleanup
+    data_callback = None
     kit.close()