+++ source/extensions/isaacsim.examples.ui/docs/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [1.1.10] - 2026-01-06
+### Changed
+- Migrate more events to Events 2.0.
+
 ## [1.1.9] - 2025-12-01
 ### Changed
 - Rename startup.py to test_startup.py